requirements.txt
-i https://pypi.org/simple
aioredis==1.3.1
asgiref==3.3.1; python_version >= '3.5'
async-timeout==3.0.1; python_full_version >= '3.5.3'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
django==3.1.7
djangorestframework==3.12.2
drf-spectacular==0.13.2
gunicorn==20.0.4
h11==0.12.0; python_version >= '3.6'
hiredis==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==3.4.0; python_version < '3.8'
inflection==0.5.1; python_version >= '3.5'
iso639-lang==0.0.9
jsonschema==3.2.0
pyrsistent==0.17.3; python_version >= '3.5'
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
redis==3.5.3
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
typing-extensions==3.7.4.3; python_version < '3.8'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
uvicorn==0.13.3
zipp==3.4.0; python_version >= '3.6'
